Pull all the various configset zipping methods into one helper.
